src/requirements.txt
altgraph==0.17.2
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
docutils==0.16
future==0.18.2
idna==2.10
importlib-metadata==3.7.2
keyring==23.0.0
packaging==20.9
pefile==2021.9.3
Pillow==9.0.1
pycryptodome==3.14.0
Pygments==2.8.1
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.1
pyparsing==2.4.7
pywin32-ctypes==0.2.0
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
tqdm==4.59.0
ttkbootstrap==1.7.1
twine==3.3.0
urllib3==1.26.3
webencodings==0.5.1
wincertstore==0.2
zipp==3.4.1
